Tag選択のメソッドをRssFeedTypeに移譲
